requirements.txt
fastapi
uvicorn
openai
Pillow
transformers
python-multipart
faiss-cpu
torch
torchvision
torchaudio